+++ src/test/java/kitchenpos/order/application/OrderValidatorTest.java
@@ -13,7 +13,7 @@ import kitchenpos.menu.domain.MenuProducts;
 import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.dto.OrderLineItemRequest;
 import kitchenpos.order.dto.OrderRequest;
-import kitchenpos.product.domain.Product;
+import kitchenpos.menu.domain.Product;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.domain.OrderTableRepository;
 import org.assertj.core.api.ThrowableAssert;
